Move database functions to one file
