+++ backend/src/httpServer.ts
@@ -3,15 +3,14 @@ import servers from "./server";
 import usersRouter from "./routes/usersRoute";
 import authRouter from "./routes/authRoute";
 import chatRouter from "./routes/chatRoute";
-import importInitialData from "./data/importData";
-import { cleanUpData } from "./db";
+import { cleanUpData, importInitialData } from "./db";
 
 const { app } = servers;
 
 dotenv.config();
 
-importInitialData().then((res) => console.log(res));
 cleanUpData();
+importInitialData().then((res) => console.log(res));
 
 app.use("/users", usersRouter);
 app.use("/auth", authRouter);
